requirements.txt
crewai
crewai[tools]
load_dotenv
#python-dotenv
langchain
langchain-openai
langchain-huggingface
langchain_community